+++ pages/_app.page.tsx
@@ -14,6 +14,7 @@ import { BorrowModal } from 'src/components/transactions/Borrow/BorrowModal';
 import { ClaimRewardsModal } from 'src/components/transactions/ClaimRewards/ClaimRewardsModal';
 import { CollateralChangeModal } from 'src/components/transactions/CollateralChange/CollateralChangeModal';
 import { EmodeModal } from 'src/components/transactions/Emode/EmodeModal';
+import { CaptchaFaucetModal } from 'src/components/transactions/Faucet/CaptchaFaucetModal';
 import { FaucetModal } from 'src/components/transactions/Faucet/FaucetModal';
 import { GasStationProvider } from 'src/components/transactions/GasStation/GasStationProvider';
 import { RateSwitchModal } from 'src/components/transactions/RateSwitch/RateSwitchModal';
@@ -85,6 +86,7 @@ export default function MyApp(props: MyAppProps) {
                           <EmodeModal />
                           <SwapModal />
                           <FaucetModal />
+                          <CaptchaFaucetModal />
                         </GasStationProvider>
                       </AppDataProvider>
                     </BackgroundDataProvider>
